Scanner refactoring and tests, handler structure

+++ syncer/scanner.py
@@ -1,7 +1,7 @@
 import sys
 import logging
 from pathlib import Path
-from hashlib import md5, file_digest
+from hashlib import sha1, md5, file_digest
 from enum import StrEnum
 from dataclasses import dataclass
 from pprint import pprint
@@ -10,8 +10,11 @@ log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
 
 # @TODO:
-# determine_action(source_file, target_file) -> Action. (dry-run problem)
 # file actions in try/except w/ logging
+
+class Strategy(StrEnum):
+    STATS = 'stats'
+    HASH = 'hash'
 
 class Action(StrEnum):
     COPY = 'copy'
@@ -23,51 +26,54 @@ class Diff:
     target: str | None
     Action: Action
 
-class Comparator:
+class Scanner:
     def __str__(self):
         return 'Custom class that gets a source and target directory and yields Diff objects.'
 
     def __repr__(self):
-        return f'Comparator with {len(self._source_set)=} and {len(self._target_set)=} items.'
+        return f'Scanner with {len(self.source_set)=} and {len(self.target_set)=} items.'
 
-    def __init__(self, source: Path, target: Path, strategy='stats'):
-        self._source_dir = source
-        self._target_dir = target
-        self.strategy = strategy
-        self._source_set = Comparator.get_filelist_in_dir(source)
-        self._target_set = Comparator.get_filelist_in_dir(target)
+    def __init__(self, source: Path, target: Path, strategy=Strategy.STATS):
+        self.source_dir = source
+        self.target_dir = target
+        self.source_set = Scanner.get_filelist_in_dir(source)
+        self.target_set = Scanner.get_filelist_in_dir(target)
+        strategy_map = {
+            Strategy.STATS: Scanner.get_stats,
+            Strategy.HASH: Scanner.get_hash,
+        }
+        self.comparison_fn = strategy_map[strategy]
 
     def __len__(self):
-        return len(self._source_set)
+        return len(self.source_set)
 
     def run(self):
-        deletions = self._target_set - self._source_set
-        additions = self._source_set - self._target_set
-        intersection = self._source_set & self._target_set
+        deletions = self.target_set - self.source_set
+        additions = self.source_set - self.target_set
+        intersection = self.source_set & self.target_set
 
         for item in sorted(deletions, reverse=True): # need to delete folder contents first
-            abspath = self._target_dir / item
+            abspath = self.target_dir / item
             yield Diff(source=None, target=str(abspath), Action=Action.DELETE)
 
         for item in sorted(additions):
             if item.parent not in additions: # whole directory to be copied, skip sub-tree
-                abspath_source = self._source_dir / item
-                abspath_target = self._target_dir / item
+                abspath_source = self.source_dir / item
+                abspath_target = self.target_dir / item
                 yield Diff(source=str(abspath_source), target=str(abspath_target), Action=Action.COPY)
 
         for item in sorted(intersection):
-            abspath_source = self._source_dir / item
-            abspath_target = self._target_dir / item
-            if not self.compare_files(abspath_source, abspath_target): # shutil.copy replaces existing
-                yield Diff(source=str(abspath_source), target=str(abspath_target), Action=Action.COPY)
+            abspath_source = self.source_dir / item
+            abspath_target = self.target_dir / item
+            if abspath_source.is_file() or abspath_target.is_file():
+                if not self.compare_files(abspath_source, abspath_target): # shutil.copy replaces existing
+                    yield Diff(source=str(abspath_source), target=str(abspath_target), Action=Action.COPY)
 
     def compare_files(self, sourcefile, targetfile) -> bool:
         try:
-            comparison_fn = getattr(Comparator, f'get_{self.strategy}')
+            return self.comparison_fn(sourcefile) == self.comparison_fn(targetfile)
         except AttributeError as exc:
             raise NotImplementedError(exc) from None
-        else:
-            return comparison_fn(sourcefile) == comparison_fn(targetfile)
     
 
     # get file metadata for comparison
@@ -77,9 +83,9 @@ class Comparator:
         return stats.st_size, stats.st_mtime
 
     @staticmethod
-    def get_hash(filepath: Path | str) -> str:
+    def get_hash(filepath: Path | str, algo: str='md5') -> str:
         with open(filepath, 'rb') as file:
-            checksum = file_digest(file, 'md5').hexdigest()
+            checksum = file_digest(file, algo).hexdigest()
         return checksum
 
     @staticmethod
@@ -99,7 +105,7 @@ class Comparator:
 # WindowsPath('Screenshot 2025-11-11 190933.png'), WindowsPath('Screenshot 2025-11-11 225208.png')}
 # target = {WindowsPath('deldir'), WindowsPath('deldir/1'), WindowsPath('deldir/delfile_indir.txt'), 
 # WindowsPath('Screenshot 2025-11-11 190933.png'), WindowsPath('Screenshot 2025-11-11 225208.png'), WindowsPath('yeah')} 
-# c = Comparator(source, target)
+# c = Scanner(source, target)
 # print(c)
 
 def info():
